+++ CHANGELOG.md
@@ -13,6 +13,8 @@ After this we will switch probably to real [Semantic Versioning 2.0.0](http://se
 ###3.5.41
 * Feature 1032: Improvements of the Vert.x Generator and enrichers
 * Removed unused Maven goals. Please contact us if something's missing for you.
+* Fix 1276: Proper inclusion of webapp's war regardless of the final name
+* Feature: New 'path' config option for the webapp generator to set the context path
 
 ###3.5.40
 * Feature 1264: Added `osio` profile, with enricher to apply OpenShift.io space labels to resources